Roles

+++ controllers/categories.js
@@ -1,10 +1,11 @@
+'use strict';
+
 const { validationResult }  = require( 'express-validator/check' );
 const repository = require( '../repositories/category' );
 
 exports.index = async ( req, res, next ) => {
     try {
-        const categories = await repository.allByUserId( req.userId );
-        //console.log( categories );
+        const categories = await repository.allByUserId( req.params.user_id );
         res.status(200).json(categories);
     }
     catch ( e ) {
@@ -12,17 +13,39 @@ exports.index = async ( req, res, next ) => {
     }
 };
 
+exports.show = async ( req, res, next ) => {
+        const id        = parseInt(req.params.category_id, 10);
+        const role      = req.role;
+        const UserId    = parseInt(req.params.user_id, 10);
+
+        const resource = await repository.findOne( { id } );
+
+        if ( ! resource ) 
+            return res.status( 400 ).json({ message: "Categoria não encontrada" });
+
+        if ( resource.UserId != UserId && role !== 'admin' ) 
+            return res.status( 403 ).json({ message: "Você não tem permissão para isso." });
+    
+        return res.status(200).json( resource );
+};
+
 exports.create = async ( req, res, next ) => {
         try {
-            const { label, color } = req.body;
+            const UserId    = req.params.user_id;
             const errors    = validationResult( req );
+            const role      = req.role;
+            
+            const { label, color }      = req.body;
+            const authenticatedUserId   = req.userId;
 
             if ( ! errors.isEmpty() ) 
                 return res.status( 422 ).json({ errors: errors.array() });
 
+            if ( authenticatedUserId != UserId && role !== 'admin' ) 
+                return res.status( 403 ).json({ message: "Você não tem permissão para isso." });
+
             var category = await repository.create({ label, color, UserId: req.userId });
             res.status(201).send({ message: 'Categoria criada com sucesso', data: category });
-
         }
         catch ( e ) {
             console.log( 'awui' );
@@ -30,28 +53,25 @@ exports.create = async ( req, res, next ) => {
         }
 };
 
-exports.show = async ( req, res, next ) => {
-    try {
-        var category = await repository.findOne({ id: parseInt(req.params.category_id, 10), UserId: req.userId });
-        
-        if ( category ) {
-            return res.status(200).json({ id: category.id, label: category.label, color: category.color });
-        }
-        return res.status(200).json(null);
-    }
-    catch ( e ) {
-        console.log( e );
-        return res.status( 500 ).json({ message: 'Erro localizar categoria', error: e });
-    }
-};
-
 exports.update = async ( req, res, next ) => {
     try {
-        const { label, color } = req.body;
-        const where     = { UserId: req.params.user_id, id: req.params.category_id };
+        const id        = req.params.category_id;
+        const UserId    = req.params.user_id;
+        const where     = { UserId, id };
+        const role      = req.role;
+
+        const { label, color }  = req.body;
+        const resource          = await repository.findOne( { id } );
+
+        if ( ! resource ) 
+            return res.status( 400 ).json({ message: "Categoria não encontrada." });
+
+        if ( resource.UserId != UserId && role !== 'admin' ) 
+            return res.status( 403 ).json({ message: "Você não tem permissão para isso." });
         
-        await repository.update( { label, color }, { where: where } );    
-        return res.status(200).json({ message: "Categoria atualizada com sucesso" });
+        await repository.update( { label, color }, { where: where } );  
+         
+        return res.status( 200 ).json({ message: "Categoria atualizada com sucesso" });
     }
     catch ( e ) {
         console.log( e );
@@ -61,13 +81,23 @@ exports.update = async ( req, res, next ) => {
 
 exports.destroy = async ( req, res, next ) => {
     try {
-        const where     = { UserId: req.params.user_id, id: req.params.category_id };
+        const id            = parseInt(req.params.category_id, 10);
+        const UserId        = parseInt(req.params.user_id, 10);
+        const role          = req.role;
+        const where         = { UserId, id };
+    
+        const resource = await repository.findOne( { id } );
+    
+        if ( ! resource ) 
+            return res.status( 400 ).json({ message: "Categoria não encontrada." });
+
+        if ( resource.UserId != UserId && role !== 'admin' ) 
+            return res.status( 403 ).json({ message: "Você não tem permissão para isso." })
         
         await repository.destroy( { where: where } );    
         return res.status(200).json({ message: "Categoria removida com sucesso." });
     }
     catch ( e ) {
-        console.log( e );
         return res.status( 500 ).json({ message: 'Erro ao remover categoria', error: e });
     }
 }
